requirements.txt
aiobotocore==2.13.0
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
arp-scan==2020.8.15
attrs==23.2.0
botocore==1.34.106
cachetools==5.3.3
capstone==5.0.1
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.6.0
gcsfs==2024.6.0
google-api-core==2.19.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
idna==3.6
jmespath==1.0.1
kiwisolver==1.4.5
leechcorepyc==2.18.6
matplotlib==3.8.4
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pefile==2023.2.7
pillow==10.2.0
prettytable==3.10.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
pyfiglet==1.0.2
pyparsing==3.1.2
pyping==0.0.6
python-dateutil==2.9.0.post0
python-nmap==0.7.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
s3fs==2024.6.0
scapy==2.5.0
six==1.16.0
urllib3==2.2.1
wcwidth==0.2.13
wrapt==1.16.0
yara-python==4.5.1
yarl==1.9.4
